Thomas renders in the center of the map

+++ core/src/za/ttd/InGameObjects/BadBreath.java
@@ -3,9 +3,6 @@ package za.ttd.InGameObjects;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
 
-/**
- * Created by s213391244 on 7/24/2015.
- */
 public class BadBreath extends Enemy {
     static final int score = 100;
 
@@ -25,12 +22,12 @@ public class BadBreath extends Enemy {
     }
 
     @Override
-    public int getX() {
+    public float getX() {
         return 0;
     }
 
     @Override
-    public int getY() {
+    public float getY() {
         return 0;
     }
 }
